src/assignments/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
attrs==22.1.0
    # via jsonschema
genson==1.2.2
    # via -r requirements.in
importlib-metadata==4.12.0
    # via jsonschema
importlib-resources==5.9.0
    # via jsonschema
jsonschema==4.13.0
    # via -r requirements.in
pkgutil-resolve-name==1.3.10
    # via jsonschema
pyrsistent==0.18.1
    # via jsonschema
ruamel-yaml==0.17.21
    # via -r requirements.in
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
typing-extensions==4.3.0
    # via
    #   importlib-metadata
    #   jsonschema
zipp==3.8.1
    # via
    #   importlib-metadata
    #   importlib-resources
